requirements.txt
Pillow
pymem
dclimplode
numpy
build
opencv-python
pyinstaller